+++ idea/frontend/src/features/voucher/ui/index.ts
@@ -1,15 +1,7 @@
 import { IssueVoucher } from './issue-voucher';
 import { ProgramVoucherSelect, CodeVoucherSelect } from './voucher-select';
-import { VoucherTable, VoucherTableDeprecated } from './voucher-table';
+import { VoucherTable } from './voucher-table';
 import { VoucherBadge } from './voucher-badge';
-import { UseVoucherCheckboxDeprecated } from './use-voucher-checkbox-deprecated';
+import { VoucherCard } from './voucher-card';
 
-export {
-  IssueVoucher,
-  ProgramVoucherSelect,
-  CodeVoucherSelect,
-  VoucherTable,
-  VoucherBadge,
-  UseVoucherCheckboxDeprecated,
-  VoucherTableDeprecated,
-};
+export { IssueVoucher, ProgramVoucherSelect, CodeVoucherSelect, VoucherTable, VoucherBadge, VoucherCard };