+++ idea/frontend/src/features/voucher/ui/issue-voucher/issue-voucher.tsx
@@ -1,37 +1,37 @@
 import { HexString } from '@gear-js/api';
 import { Button } from '@gear-js/ui';
-import { useState } from 'react';
 import cx from 'clsx';
 
 import { withAccount } from '@/shared/ui';
-import actionLinkStyles from '@/shared/ui/ActionLink/ActionLink.module.scss';
 
 import CouponSVG from '../../assets/coupon.svg?react';
 import { IssueVoucherModal } from '../issue-voucher-modal';
+import { useModal } from '../../hooks';
+import styles from './issue-voucher.module.scss';
 
 type Props = {
-  programId: HexString;
-  transparent?: boolean;
+  programId?: HexString;
+  buttonSize?: 'small' | 'medium';
+  buttonColor?: 'secondary' | 'light' | 'transparent';
+  onSubmit?: () => void;
 };
 
-const IssueVoucher = withAccount(({ programId, transparent }: Props) => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const openModal = () => setIsModalOpen(true);
-  const closeModal = () => setIsModalOpen(false);
+const IssueVoucher = withAccount(({ programId, buttonColor = 'light', buttonSize = 'medium', onSubmit }: Props) => {
+  const [isModalOpen, openModal, closeModal] = useModal();
 
   return (
     <>
       <Button
         icon={CouponSVG}
         text="Create Voucher"
-        color={transparent ? 'transparent' : 'secondary'}
-        size="small"
-        className={cx(transparent && actionLinkStyles.link)}
-        noWrap
+        size={buttonSize}
+        color={buttonColor}
+        className={cx(buttonColor === 'transparent' && styles.link)}
         onClick={openModal}
+        noWrap
       />
 
-      {isModalOpen && <IssueVoucherModal programId={programId} close={closeModal} />}
+      {isModalOpen && <IssueVoucherModal programId={programId} close={closeModal} onSubmit={onSubmit} />}
     </>
   );
 });