+++ idea/frontend/src/shared/ui/form/index.ts
@@ -1,9 +1,9 @@
 import formStyles from './Form.module.scss';
 import { Input } from './input';
 import { ValueField } from './value-field';
-import { Form } from './form';
 import { Checkbox } from './checkbox';
 import { Radio } from './radio';
 import { Select } from './select';
+import { SearchForm } from './search-form';
 
-export { formStyles, Input, ValueField, Form, Checkbox, Radio, Select };
+export { formStyles, Input, ValueField, Checkbox, Radio, Select, SearchForm };