+++ api/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.37.2
+
+_04/11/2024_
+
+### Changes
+https://github.com/gear-tech/gear-js/pull/1528
+- Subscribe to blocks and gear events starting from a particular block
+
+## 0.37.1
+
+_04/10/2024_
+
+### Changes
+https://github.com/gear-tech/gear-js/pull/1522
+- Subsribe to gear events only on finalized blocks
+
 ## 0.37.0
 
 _03/12/2024_